combine NoPools & PoweredOffPools diagnostics
