+++ TestDs3/Helpers/Diagnostics/TestPoolsDiagnostic.cs
@@ -23,8 +23,30 @@ using NUnit.Framework;
 namespace TestDs3.Helpers.Diagnostics
 {
     [TestFixture]
-    public class TestPoweredOffPools
+    public class TestPoolsDiagnostic
     {
+        [Test]
+        public void TestGetWithNoPools()
+        {
+            var client = new Mock<IDs3Client>(MockBehavior.Strict);
+            client
+                .Setup(c => c.GetPoolsSpectraS3(It.IsAny<GetPoolsSpectraS3Request>()))
+                .Returns(DiagnosticsStubResponses.NoPools);
+
+            var noPools = new PoolsDiagnostic();
+            var ds3DiagnosticClient = new Ds3DiagnosticClient
+            {
+                Client = client.Object
+            };
+            var noPoolsResult = noPools.Get(ds3DiagnosticClient);
+
+            Assert.AreEqual(Ds3DiagnosticsCode.NoPoolsFound, noPoolsResult.ClientResult.Code);
+            Assert.AreEqual(DiagnosticsMessages.NoPoolsFound, noPoolsResult.ClientResult.ErrorMessage);
+            Assert.AreEqual(null, noPoolsResult.ClientResult.ErrorInfo);
+
+            client.VerifyAll();
+        }
+
         [Test]
         public void TestGetWithNoPoweredOffPools()
         {
@@ -33,7 +55,7 @@ namespace TestDs3.Helpers.Diagnostics
                 .Setup(c => c.GetPoolsSpectraS3(It.IsAny<GetPoolsSpectraS3Request>()))
                 .Returns(DiagnosticsStubResponses.NoPoweredOffPools);
 
-            var poweredOffPools = new PoweredOffPoolsDiagnostic();
+            var poweredOffPools = new PoolsDiagnostic();
             var ds3DiagnosticClient = new Ds3DiagnosticClient
             {
                 Client = client.Object
@@ -56,7 +78,7 @@ namespace TestDs3.Helpers.Diagnostics
                 .Setup(c => c.GetPoolsSpectraS3(It.IsAny<GetPoolsSpectraS3Request>()))
                 .Returns(DiagnosticsStubResponses.OnePoweredOffPool);
 
-            var poweredOffPools = new PoweredOffPoolsDiagnostic();
+            var poweredOffPools = new PoolsDiagnostic();
             var ds3DiagnosticClient = new Ds3DiagnosticClient
             {
                 Client = client.Object
@@ -78,7 +100,7 @@ namespace TestDs3.Helpers.Diagnostics
                 .Setup(c => c.GetPoolsSpectraS3(It.IsAny<GetPoolsSpectraS3Request>()))
                 .Returns(DiagnosticsStubResponses.TwoPoweredOffPool);
 
-            var poweredOffPools = new PoweredOffPoolsDiagnostic();
+            var poweredOffPools = new PoolsDiagnostic();
             var ds3DiagnosticClient = new Ds3DiagnosticClient
             {
                 Client = client.Object
