+++ TestDs3/TestDs3.csproj
@@ -70,9 +70,8 @@
     <Compile Include="Helpers\Diagnostics\DiagnosticsStubResponses.cs" />
     <Compile Include="Helpers\Diagnostics\TestCacheNearCapacity.cs" />
     <Compile Include="Helpers\Diagnostics\TestDs3Diagnostic.cs" />
-    <Compile Include="Helpers\Diagnostics\TestNoPoolsDiagnostic.cs" />
+    <Compile Include="Helpers\Diagnostics\TestPoolsDiagnostic.cs" />
     <Compile Include="Helpers\Diagnostics\TestTapesDiagnostic.cs" />
-    <Compile Include="Helpers\Diagnostics\TestPoweredOffPools.cs" />
     <Compile Include="Helpers\Diagnostics\TestReadFromTapeDiagnostic.cs" />
     <Compile Include="Helpers\Diagnostics\TestWriteToTapeDiagnostic.cs" />
     <Compile Include="Helpers\JobResponseStubs.cs" />
